Refactored map card component
Added spacing between multiple card elements and pulled the counting button component out into its own file.
